refactor: split the boot of application for simplify testing
