+++ package.json
@@ -2,11 +2,11 @@
   "name": "busbud-coding-challenge",
   "version": "1.0.0",
   "description": "Busbud backend coding challenge",
-  "main": "src/boot.js",
+  "main": "src/run.js",
   "scripts": {
-    "start": "node src/boot.js",
-    "start.dev": "nodemon src/boot.js",
-    "test": "nyc mocha",
+    "start": "node src/run.js",
+    "start.dev": "nodemon src/run.js",
+    "test": "nodemon --exec 'nyc mocha'",
     "coverage": "nyc report --reporter=text-lcov > coverage.lcov",
     "docker-start": "docker-compose up",
     "lint": "eslint ./",
