+++ src/run.js
@@ -1,0 +1,5 @@
+const boot = require('./boot')
+
+;(async () => {
+  await boot()
+})()
